pyproject.toml
[project]
name = "dolarbcvwidget"
version = "0.1.0"
description = ""
authors = [
    {name = "OriC28",email = "orianacolina.perea@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.4",
    "pillow>=10.0.0"
]

[tool.poetry]
packages = [{include = "dolarbcvwidget", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
